Align test package names with main code at com.pambrose.srcref
