+++ src/test/kotlin/com/pambrose/srcref/core/ContentCacheTest.kt
@@ -14,7 +14,7 @@
  *   limitations under the License.
  */
 
-package com.srcref.core
+package com.pambrose.srcref.core
 
 import com.pambrose.srcref.ContentCache
 import com.pambrose.srcref.ContentCache.CacheContent
